script/requirements.txt
openai>=1.97.1
python-dotenv>=1.1.1
